move provider definitions to a single file
