+++ .env.example
@@ -7,7 +7,7 @@ METRIQ_CLIENT_API_KEY="<METRIQ_CLIENT_API_KEY>"
 
 # IBM
 # Obtained at: https://quantum.ibm.com
-IBM_QISKIT_API_KEY="<IBM_QISKIT_API_KEY>"
+QISKIT_IBM_TOKEN="<QISKIT_IBM_TOKEN>"
 
 # Quantinuum
 # Obtained at: https://um.qapi.quantinuum.com
