+++ metriq_gym/provider.py
@@ -1,6 +1,17 @@
+from enum import StrEnum
+
 from qbraid.runtime import QiskitRuntimeProvider
 from qbraid.runtime import IonQProvider
-from metriq_gym.providers.provider import ProviderType
+
+
+class ProviderType(StrEnum):
+    IBMQ = "ibmq"
+    IONQ = "ionq"
+
+    @classmethod
+    def list(cls):
+        return [provider.value for provider in cls]
+
 
 PROVIDERS = {
     ProviderType.IBMQ: QiskitRuntimeProvider,
