+++ pyproject.toml
@@ -29,6 +29,7 @@ qiskit-ionq = "^0.5.12"
 scipy = "^1.15.0"
 tabulate = "^0.9.0"
 qbraid = {extras = ["ionq"], version = "^0.9.0"}
+qiskit-qasm3-import = "^0.5.1"
 
 [tool.poetry.group.dev.dependencies]
 mypy = "^1.14.1"
